+++ lib/kubereq/pod_exec.ex
@@ -75,6 +75,7 @@ defmodule Kubereq.PodExec do
   def start_link(args) do
     {into, args} = Keyword.pop!(args, :into)
     {req, args} = Keyword.pop!(args, :req)
+    {genserver_opts, args} = Keyword.pop(args, :genserver_opts, [])
 
     opts =
       args
@@ -83,7 +84,16 @@ defmodule Kubereq.PodExec do
 
     req = Req.merge(req, opts)
 
-    Kubereq.Connect.start_link(__MODULE__, req, %{into: into})
+    Kubereq.Connect.start_link(__MODULE__, req, %{into: into}, genserver_opts)
+  end
+
+  def child_spec(init_arg) do
+    default = %{
+      id: __MODULE__,
+      start: {Kubereq.PodExec, :start_link, [init_arg]}
+    }
+
+    Supervisor.child_spec(default, [])
   end
 
   @doc """