+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ### Added | Changed | Deprecated | Removed | Fixed | Security -->
 
+### Fixed
+
+* `child_spec` for `Kubereq.PodLogs` and `Kubereq.PodExec`
+
 <!--------------------- Don't add new entries after this line --------------------->
 
 ## 0.4.1 - 2024-12-26